requirements.txt
autopep8
flake8
py-zabbix
fire
